camera modal clean up
